working on split to router files
